requirements.txt
dotenv
argparse
google.generativeai